Column length check test
This test will produce warnings if values in a configurable set of
columns exceed a specified length. New columns could be added to the
test by editing the test and setting the new column's max length, and
desired action (warning or error).

Current columns checked:
- culture: max length 6 results in warning if exceeded.
- language: max length 6 results in warning if exceeded.
- script: max length 6 results in warning if exceeded.

Also removed all trailing periods for results that ended with a numeric
value or a list.

Also removed unnecessary checks on array_key_exists().